[#677] Separate routes files for better organization
* Routes are split into multiple files: auth, account, product, secrets, and public.
* Improves maintainability by logically grouping related routes.
* Auth routes include a logout handler that clears local storage.
* Ensures that unauthenticated users are redirected to login if needed.
